Move ItemsContainer and MoonsContainer to a containers folder.
