+++ GuysNight.LethalCompanyMod.BalancedItems/Containers/MoonsContainer.cs
@@ -1,7 +1,7 @@
 ﻿using GuysNight.LethalCompanyMod.BalancedItems.Models.Moons;
 using System.Collections.Generic;
 
-namespace GuysNight.LethalCompanyMod.BalancedItems {
+namespace GuysNight.LethalCompanyMod.BalancedItems.Containers {
 	internal static class MoonsContainer {
 		internal static Dictionary<string, MoonProperties> Moons { get; } = new Dictionary<string, MoonProperties>();
 	}
