+++ GuysNight.LethalCompanyMod.BalancedItems/Patches/GrabbableObjectPatches.cs
@@ -1,5 +1,6 @@
 ﻿#pragma warning disable	S1118
 
+using GuysNight.LethalCompanyMod.BalancedItems.Containers;
 using GuysNight.LethalCompanyMod.BalancedItems.Models.Items;
 using GuysNight.LethalCompanyMod.BalancedItems.Utilities;
 using HarmonyLib;
