+++ GuysNight.LethalCompanyMod.BalancedItems/Patches/RoundManagerPatches.cs
@@ -1,5 +1,6 @@
 ﻿#pragma warning disable	S1118
 
+using GuysNight.LethalCompanyMod.BalancedItems.Containers;
 using GuysNight.LethalCompanyMod.BalancedItems.Utilities;
 using HarmonyLib;
